Update
now this adds the HLK-PM01 so that this can have it's own AC-DC adapter built in, and has way more comfortable connection cable management inside a box. Next iterations will be replacing the connectors for the SDA and SCL (and all those 7 signal pins) for more robust and compact ones. In this iteration the sizing of the AC traces were all changed to match a 3mm width which allows for 5A input current handling at most, with a delta of 10°C.